Fixing remote import error introduced in e3311d3
