+++ circusort/__init__.py
@@ -6,7 +6,7 @@ from . import obj
 from . import plt
 
 from .base import create_director
-from .utils import *
+from .base import utils
 
 
 __version__ = '0.0.1'
